Move stuff around in circles.
