+++ src/main/java/ch/njol/skript/lang/function/FunctionReference.java
@@ -33,7 +33,7 @@ import ch.njol.skript.util.LiteralUtils;
 import ch.njol.util.StringUtils;
 import org.bukkit.event.Event;
 import org.eclipse.jdt.annotation.Nullable;
-import org.skriptlang.skript.lang.function.Contract;
+import org.skriptlang.skript.lang.util.Contract;
 
 import java.util.ArrayList;
 import java.util.Arrays;
